# assets/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Shopping Basket API","description":"Shopping Basket RESTful API"},"paths":{"/api/v1/Basket":{"post":{"tags":["Basket"],"operationId":"CreateBasket","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/CreateBasketRequest"}}],"responses":{"200":{"description":"Success"}}}},"/api/v1/Basket/{basketId}":{"get":{"tags":["Basket"],"operationId":"GetBasket","consumes":[],"produces":["application/json"],"parameters":[{"name":"basketId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/BasketResponse"}}}},"put":{"tags":["Basket"],"operationId":"UpdateItemQuantity","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["application/json"],"parameters":[{"name":"basketId","in":"path","required":true,"type":"string"},{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/AdjustQuantityRequest"}}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Basket"],"operationId":"AddItem","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["application/json"],"parameters":[{"name":"basketId","in":"path","required":true,"type":"string"},{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/AddItemRequest"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Basket"],"operationId":"RemoveItem","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["application/json"],"parameters":[{"name":"basketId","in":"path","required":true,"type":"string"},{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/RemoveItemRequest"}}],"responses":{"200":{"description":"Success"}}}},"/api/v1/Basket/{basketId}/clear":{"delete":{"tags":["Basket"],"operationId":"EmptyBasket","consumes":[],"produces":["application/json"],"parameters":[{"name":"basketId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"CreateBasketRequest":{"type":"object","properties":{"customerId":{"format":"uuid","type":"string"}}},"AddItemRequest":{"type":"object","properties":{"productName":{"type":"string"},"price":{"format":"double","type":"number"},"quantity":{"format":"int32","type":"integer"}}},"AdjustQuantityRequest":{"type":"object","properties":{"productName":{"type":"string"},"quantity":{"format":"int32","type":"integer"}}},"RemoveItemRequest":{"type":"object","properties":{"productName":{"type":"string"}}},"BasketResponse":{"type":"object","properties":{"basketId":{"type":"string"},"customerId":{"format":"uuid","type":"string"},"totalPrice":{"format":"double","type":"number"},"totalItems":{"format":"int32","type":"integer"},"items":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/BasketItemResponse"}}}},"BasketItemResponse":{"type":"object","properties":{"productName":{"type":"string"},"price":{"format":"double","type":"number"},"quantity":{"format":"int32","type":"integer"}}}}}